changed look of home page
