+++ frontend/src/ConfigEditor.jsx
@@ -13,12 +13,18 @@ import DomTree from "./components/DomTree.jsx";
 import NodeSelection from "./components/NodeSelection.jsx";
 import { useConfigs } from "./context/ConfigContext.jsx";
 import { useRetrievalInstructions } from "./context/RetrievalInstructionContext.jsx";
+import { useTheme } from "@mui/material/styles";
 
 function ConfigEditor({ placeholderRoot }) {
+  const theme = useTheme();
   const location = useLocation();
   const { postConfig } = useConfigs();
 
   const {
+    url,
+    setUrl,
+    setTree,
+    lastBuiltUrlRef,
     name,
     setName,
     description,
@@ -34,6 +40,7 @@ function ConfigEditor({ placeholderRoot }) {
     if (path === "/configs/new") {
       setName(null);
       setDescription(null);
+      setUrl("");
       clearInstructions();
       return;
     }
@@ -44,6 +51,7 @@ function ConfigEditor({ placeholderRoot }) {
       if (cfg) {
         setName(cfg.name ?? null);
         setDescription(cfg.description ?? null);
+        setUrl(cfg.url ?? "");
         setRetrievalInstructions(
           Array.isArray(cfg.retrieval_instructions)
             ? [...cfg.retrieval_instructions]
@@ -52,10 +60,44 @@ function ConfigEditor({ placeholderRoot }) {
       } else {
         setName(null);
         setDescription(null);
+        setUrl("");
         clearInstructions();
       }
     }
-  }, [location.pathname]); // <- only depend on pathname
+  }, [location.pathname]);
+
+  const buildTree = async (givenUrl) => {
+    const targetUrl = givenUrl ?? url;
+    if (!targetUrl) return;
+
+    try {
+      const res = await fetch(
+        import.meta.env.VITE_API_URL + "/api/v1/scraper/dom-tree/build",
+        {
+          method: "POST",
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ url: targetUrl }),
+        }
+      );
+
+      const json = await res.json();
+      setTree(json.root);
+      lastBuiltUrlRef.current = targetUrl;
+    } catch (err) {
+      console.error("buildTree error:", err);
+    }
+  };
+
+  useEffect(() => {
+    if (url === lastBuiltUrlRef.current) return;
+    if (!url) return;
+
+    const t = setTimeout(() => buildTree(url), 700);
+    return () => clearTimeout(t);
+  }, [url]);
 
   const saveAndContinue = async () => {
     await postConfig(name ?? "", description ?? "", retrievalInstructions);
@@ -74,6 +116,42 @@ function ConfigEditor({ placeholderRoot }) {
         p: 2,
       }}
     >
+      <h2
+        style={{
+          fontSize: "1.5rem",
+          fontWeight: 600,
+          color: theme.palette.text.primary,
+          margin: 0,
+        }}
+      >
+        New Config
+      </h2>
+
+      <Paper sx={{ p: { xs: 2.5, md: 3 }, mb: 2 }}>
+        <Stack spacing={1} alignItems="center" textAlign="center">
+          <Typography variant="h6">Website URL</Typography>
+
+          <Box sx={{ width: "100%", maxWidth: 760, mx: "auto" }}>
+            <TextField
+              fullWidth
+              label="Website URL"
+              placeholder="https://example.com"
+              value={url ?? ""}
+              onChange={(e) => setUrl(e.target.value)}
+              slotProps={{
+                input: {
+                  sx: {
+                    fontSize: 16,
+                    height: 52,
+                    "& .MuiInputBase-input": { py: 1, lineHeight: 1.5 },
+                  },
+                },
+              }}
+            />
+          </Box>
+        </Stack>
+      </Paper>
+
       <Box sx={{ flex: 1, display: "flex", gap: 2, minHeight: 0 }}>
         <Box
           sx={{
