+++ frontend/src/Home.jsx
@@ -1,24 +1,34 @@
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import ConfigSource from './components/configs/ConfigSource.jsx';
-import { Box, Paper, Stack, Typography, TextField, Button } from '@mui/material';
+import {
+  Box,
+  Paper,
+  Stack,
+  Typography,
+  Button,
+  ToggleButtonGroup,
+  ToggleButton,
+} from '@mui/material';
+import UploadIcon from '@mui/icons-material/Upload';
+import LibraryBooksIcon from '@mui/icons-material/LibraryBooks';
+import AddCircleIcon from '@mui/icons-material/AddCircle';
+import ImportConfig from './components/ImportConfig.jsx';
 import { useRetrievalInstructions } from './context/RetrievalInstructionContext.jsx';
 import { useConfigs } from "./context/ConfigContext.jsx";
+import { useTheme } from "@mui/material/styles";
 
 function Home() {
+  const theme = useTheme();
   const {
     url,
-    setUrl,
-    tree,
-    setTree,
     flow,
     setFlow,
     retrievalInstructions,
-    lastBuiltUrlRef
   } = useRetrievalInstructions();
 
   const navigate = useNavigate();
   const { postConfig } = useConfigs();
+  const [importedFile, setImportedFile] = useState(null);
 
   const placeholderRoot = {
     id: 1,
@@ -27,29 +37,6 @@ function Home() {
       { id: 2, tag_type: 'h1', body: 'No Data Currently Displayed', hasData: true },
       { id: 3, tag_type: 'p', body: 'Please enter a URL', hasData: true },
     ],
-  };
-
-  const buildTree = async (givenUrl) => {
-    const targetUrl = givenUrl ?? url;
-    try {
-      const res = await fetch(
-        import.meta.env.VITE_API_URL + '/api/v1/scraper/dom-tree/build',
-        {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({ url: targetUrl })
-        }
-      );
-
-      const json = await res.json();
-      setTree(json.root);
-      lastBuiltUrlRef.current = targetUrl;
-    } catch (err) {
-      console.error('buildTree error:', err);
-    }
   };
 
   const scrapeHandler = async () => {
@@ -81,77 +68,129 @@ function Home() {
     await postConfig("My first config");
   };
 
-  useEffect(() => {
-    if (url === lastBuiltUrlRef.current) return;
-    const t = setTimeout(() => buildTree(url), 700);
-    return () => clearTimeout(t);
-  }, [url]);
+  const handleFlowChange = (_, val) => {
+    if (!val) return;
+
+    setFlow(val);
+
+    if (val === "new") {
+      navigate('/configs/new', {
+        state: { placeholderRoot: placeholderRoot }
+      });
+    }
+  };
 
   useEffect(() => {
     if (flow === "new") {
       setFlow("saved");
     }
-  }, []); 
+  }, []);
 
   return (
-    <Box sx={{ p: 1 }}>
-      <Paper sx={{ p: { xs: 2.5, md: 3 }, mb: 2.5 }}>
-        <Stack spacing={1} alignItems="center" textAlign="center">
-          <Typography variant="overline" sx={{ letterSpacing: 1.2, opacity: 0.7, lineHeight: 1.2 }}>
-            Step 1
-          </Typography>
-          <Typography variant="h6">Enter Website URL</Typography>
+    <Box
+      sx={{
+        p: 1,
+        minHeight: "calc(100vh - 250px)",
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        alignItems: "center"
+      }}
+    >
 
-          <Box sx={{ width: '100%', maxWidth: 760, mx: 'auto' }}>
-            <TextField
-              fullWidth
-              label="Website URL"
-              placeholder="https://example.com"
-              value={url}
-              onChange={(e) => setUrl(e.target.value)}
-              slotProps={{
-                input: {
-                  sx: {
-                    fontSize: 16,
-                    height: 52,
-                    '& .MuiInputBase-input': { py: 1, lineHeight: 1.5 },
-                  }
-                }
+      <h1
+        style={{
+          fontSize: "2.5rem",
+          fontWeight: 700,
+          color: theme.palette.text.primary,
+          margin: 0,
+          paddingBottom: "1rem",
+          letterSpacing: "0.5px",
+        }}
+      >
+        ScrapeGoat
+      </h1>
+
+      <Paper
+        sx={{
+          p: 2.5,
+          pb: 3,
+          mb: 2,
+          width: "100%",
+          maxWidth: 700,
+          borderRadius: 3
+        }}
+      >
+        <Stack spacing={1.8}>
+          <Stack alignItems="center" textAlign="center" spacing={0.5}>
+            <Typography variant="h6" sx={{ fontWeight: 500 }}>
+              Choose Configuration Source
+            </Typography>
+          </Stack>
+
+          <Stack direction="row" justifyContent="center">
+            <ToggleButtonGroup
+              value={flow}
+              exclusive
+              onChange={handleFlowChange}
+              size="small"
+            >
+              <ToggleButton value="new">
+                <AddCircleIcon sx={{ mr: 1 }} />
+                New Config
+              </ToggleButton>
+              <ToggleButton value="saved">
+                <LibraryBooksIcon sx={{ mr: 1 }} />
+                Saved Config
+              </ToggleButton>
+              <ToggleButton value="import">
+                <UploadIcon sx={{ mr: 1 }} />
+                Import .GOAT
+              </ToggleButton>
+            </ToggleButtonGroup>
+          </Stack>
+
+          {flow === 'saved' && (
+            <Typography
+              variant="body2"
+              sx={{
+                opacity: 0.7,
+                fontSize: '0.85rem',
+                width: '100%',
+                mx: 'auto',
+                textAlign: 'center'
               }}
+            >
+              Log in to see your saved configurations!
+            </Typography>
+          )}
+
+          {flow === 'import' && (
+            <ImportConfig
+              file={importedFile}
+              onChange={setImportedFile}
+              hint="Import an existing configuration file."
             />
-          </Box>
+          )}
         </Stack>
       </Paper>
 
-      <ConfigSource placeholderTree={placeholderRoot} />
-
-      <Paper sx={{ p: { xs: 2.5, md: 3 }, mt: 3 }}>
-        <Stack spacing={1.5} alignItems="center" textAlign="center">
-          <Typography
-            variant="overline"
-            sx={{ letterSpacing: 1.2, opacity: 0.7, lineHeight: 1.2 }}
-          >
-            Step 3
-          </Typography>
-
-          <Typography variant="h6">Run Scrape</Typography>
-
-          <Button
-            variant="contained"
-            onClick={scrapeHandler}
-            sx={{
-              mt: 1,
-              px: 3,
-              py: 1.2,
-              borderRadius: 9999,
-              fontSize: '1rem',
-              boxShadow: 6,
-            }}
-          >
-            Scrape ➔
-          </Button>
-        </Stack>
-      </Paper>
+      <Stack spacing={2} alignItems="center" textAlign="center" sx={{ mt: 2 }}>
+        <Button
+          variant="contained"
+          onClick={scrapeHandler}
+          sx={{
+            mt: 0.5,
+            px: 3,
+            py: 1.2,
+            borderRadius: 9999,
+            fontSize: '1rem',
+            boxShadow: 6,
+          }}
+        >
+          Scrape ➔
+        </Button>
+      </Stack>
     </Box>
   );
 }
